Move stream source fetcher to the saga, make some refactoring
